fix comments and failures
